back-end/requirements.txt
granian[reload]
pymongo
fastapi
web3
python-dotenv